requirements.txt
# Core deep learning stack
torch>=2.0.0
torchvision>=0.15.0

# Data handling and model utilities
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
tqdm>=4.66.0

# Image processing
Pillow>=10.2.0
opencv-python>=4.9.0.80
# OpenCV is imported as cv2 in Python code. Install package name is opencv-python, not cv2.

# Visualization and analysis
matplotlib>=3.8.0
seaborn>=0.13.0

# API backend
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.24
requests>=2.32.0
python-dotenv>=1.0.1

# Frontend ResultCard update
# No additional Python packages required.

# Optional: bulk image collection (secondary data sources)
bing-image-downloader>=1.1.2
google-genai>=0.3.0
